+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 symbo or property on right of pipe was recognized as command now.
 - add filter to loop pipe, return none insteadof empty vec.
 - add range support for loop pipe
+- add HashMap grammar
+
 
 ## [0.12.5]
 - only turn off PRINT_DIRECT in script mode.